backend/requirements.txt
fastapi
uvicorn
python-multipart
lxml
pandas
pydantic
streamlit
requests
plotly
pdfplumber
python-docx
